Move fonts to their own library
