+++ addons/fonts/src/com/BBsRs/SFUIFontsEverywhere/SFUIFontsPath.java
@@ -1,6 +1,6 @@
-package org.holoeverywhere.slider;
+package com.BBsRs.SFUIFontsEverywhere;
 
-public class SFUIDisplayFontPath {
+public class SFUIFontsPath {
 	public static String ULTRALIGHT = "fonts/SFUIDisplay-Ultralight.otf";
 	public static String LIGHT = "fonts/SFUIDisplay-Light.otf";
 	public static String MEDIUM = "fonts/SFUIDisplay-Medium.otf";
